feat: Refactor settings screen in React Native client
